Add files via upload
This repository contains a simple calculator made using HTML, CSS, and JavaScript. The calculator can perform basic mathematical operations such as addition, subtraction, multiplication, and division. It also has a dark mode theme.

The calculator is designed to be simple and easy to use. The user interface is clean and concise, and the code is well-commented. The calculator is also responsive, meaning that it will work properly on any screen size.

This project is a good starting point for anyone who wants to learn how to build a calculator using HTML, CSS, and JavaScript. It is also a good example of how to use a dark mode theme in a web application.

To use the calculator, simply enter your numbers in the input fields and click on the desired operation. The result will be displayed in the output field.

To enable dark mode, click on the "Dark Mode" button.

I hope you find this project useful!

The features of the calculator, such as the ability to perform basic mathematical operations, the dark mode theme, and the responsive design.
The target audience for the calculator, such as beginners who want to learn how to build a calculator or developers who need a simple calculator for their projects.
The technologies used to build the calculator, such as HTML, CSS, and JavaScript.
The instructions on how to use the calculator.